src/2. Models/2. LLMs/pyproject.toml
[project]
name = "2-llms"
version = "0.1.0"
description = "Basic interactions with LLMs"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.71.0",
    "langchain-anthropic>=1.0.0",
    "langchain-openai>=1.0.1",
    "langchain-ollama>=1.0.0",
    "openai>=2.6.0",
    "python-dotenv>=1.0.0",
    "streamlit>=1.50.0",
]
